moved handleStoragePermissions() to SimpleActivity.kt from MainActivity.kt introduced tryInitFileManager() to SaveAsActivity.kt removed const MANAGE_STORAGE_RC from MainActivity.kt
